% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{nr_style_file}
\alias{nr_style_file}
\title{Style single file}
\usage{
nr_style_file(path, only_spaces = TRUE)
}
\arguments{
\item{path}{Character}

\item{only_spaces}{Logical}
}
\description{
Style single file
}
